Refactoring GameManager class by creating Image reader classes.
